+++ ext/bigdecimal/math_r/exp.c
@@ -18,7 +18,7 @@
  * @raise [ArgumentError] Occurs when +prec+ is not a positive integer.
  * @raise [TypeError] Occurs when +x+ is not a numeric class.
  * @example
- *  BigMathR::EDF.exp(1, 20)
+ *  BigMathR::Exp.exp_edf(1, 20)
  *  #=> 0.27182818284590452354e1
  * @since 0.1.0
  */
@@ -38,7 +38,7 @@ __impl_exp_edf(VALUE unused_obj, VALUE x, VALUE prec)
  * @raise [ArgumentError] Occurs when +prec+ is not a positive integer.
  * @raise [TypeError] Occurs when +x+ is not a numeric class.
  * @example
- *  BigMathR::EDF.exp2(1/2r, 20)
+ *  BigMathR::Exp.exp2_edf(1/2r, 20)
  *  #=> 0.14142135623730950488e1
  * @since 0.1.0
  */
@@ -47,6 +47,46 @@ __impl_exp2_edf(VALUE unused_obj, VALUE x, VALUE prec)
 {
 	return exp2_branch(x, prec, exp2_edf);
 }
+
+/**
+ * Computes the subtracted 1 from x of the exponential function.
+ * <br>
+ * The calculation is more accurate than +exp(x)-1+ when the argument is near zero.
+ * @param x [Numeric] Numerical argument
+ * @param prec [Integer] Arbitrary precision
+ * @return [BigDecimal] Real solution
+ * @raise [ArgumentError] Occurs when +prec+ is not a positive integer.
+ * @raise [TypeError] Occurs when +x+ is not a numeric class.
+ * @example
+ *  BigMathR::Exp.expm1_identity(BigDecimal(1), 20)
+ *  #=> 0.17182818284590452354e1
+ * @since 0.2.2
+ */
+static VALUE
+__impl_expm1_identity(VALUE unused_obj, VALUE x, VALUE prec)
+{
+	return expm1_branch(x, prec, expm1_identity);
+}
+
+/**
+ * Computes the subtracted 1 from +z+ of the exponential function.
+ * 
+ * @param z [Numeric] Numerical argument
+ * @param prec [Integer] Arbitrary precision
+ * @return [Complex] Complex solution
+ * @raise [ArgumentError] Occurs when +prec+ is not a positive integer.
+ * @raise [TypeError] Occurs when +z+ is not a numeric class.
+ * @example
+ *  BigMathR::Exp.cexpm1_identity(1+1i, 20)
+ *  # => (0.46869393991588515714e0+0.22873552871788423912e1i)
+ * @since 0.2.2
+ */
+static VALUE
+__impl_cexpm1_identity(VALUE unused_obj, VALUE z, VALUE prec)
+{
+	return cexpm1_branch(z, prec, cexpm1_identity);
+}
+
 
 
 /**
@@ -62,11 +102,13 @@ __impl_exp2_edf(VALUE unused_obj, VALUE x, VALUE prec)
  *  <br>
  *  Exponential function:     +exp()+ <br>
  *  Base-2 (binary) exponent: +exp2()+ <br>
- *  
+ *  expm1 function:           +expm1()+ <br>
  */
 void
 InitVM_Exp(void)
 {
 	rb_define_module_function(rb_mExp, "exp_edf", __impl_exp_edf, 2);
 	rb_define_module_function(rb_mExp, "exp2_edf", __impl_exp2_edf, 2);
+	rb_define_module_function(rb_mExp, "expm1_identity", __impl_expm1_identity, 2);
+	rb_define_module_function(rb_mExp, "cexpm1_identity", __impl_cexpm1_identity, 2);
 }