+++ ext/bigdecimal/math_r/internal/algo/bigmath/clog2/formula.h
@@ -5,7 +5,7 @@ clog2_formula(VALUE z, VALUE prec)
 
 	rb_check_precise(prec);
 
-	w = clog_formula(z, prec);
+	w = rb_bigmath_clog(z, prec);
 	w = rb_funcall1(w, '/', rb_bigmath_const_log2(prec));
 
 	return rb_num_round(w, prec);